requirements_base.txt
# mechanize
# BeautifulSoup
# frozen versions from prod.
app-version==0.2.1
dealer==2.0.4
Django==2.0.9
django-compat==1.0.15
django-ajax-selects==1.7.1
django-appconf==1.0.1
# -e git+https://github.com/jmerdich/django-bootstrap-calendar.git@fbe22d9dd0b2c99e8a6d3a3ad8d41063e791432c#egg=django_bootstrap_calendar-master
-e git+https://github.com/rrlapointe/django-bootstrap-calendar.git@7bc68b6e14e17246530db4fcf9c25c91fe2f2d1d#egg=django_bootstrap_calendar-master
django-bootstrap-toolkit==2.15.0
#-e git+https://github.com/WPI-lnl/django-cas2.git@6220bc8ba6eb4643e3ef58b4830616f37f1a1a8f#egg=django_cas-origin_HEAD
django-compressor==1.3
django-crispy-forms==1.7.2
# -e git+https://github.com/maraujop/django-crispy-forms.git@57e9e3f9c5deb2876aa5021599671d6a65e97ed0#egg=django_crispy_forms-dev
django-debug-toolbar==1.10.1
django-extensions==2.1.3
django-formtools==2.1
django-hijack==2.1.10
django-ical==1.4
django-lineage==0.2.0
django-markdown-deux==1.0.5
django-mptt==0.9.1
django-multiupload==0.5
django-natural-duration==0.1
django-pagedown==0.1.3
django-permission==0.9.2
django-reversion==3.0.1
django-uuidfield==0.5.0
django-watson==1.5.2
docutils==0.12
html5lib==1.0b3
icalendar==3.9.1
jsmin==2.1.4
line-profiler==1.0
Markdown==2.6.2
markdown2==2.3.3
mechanize==0.2.5
Paste==2.0.2
Pillow==4.2.1
pisa==3.0.33
pyPdf==1.13
PyPDF2==1.25.1
python-dateutil==2.4.2
pytz==2015.4
raven==6.9.0
reportlab==3.2.0
six==1.11.0
sqlparse==0.2.3
tolerance==0.1.2
xhtml2pdf==0.2b1
django-cas-ng==3.5.10
ldap3==1.4.0
django-environ==0.4.3
django-polymorphic==2.0.3
